Yet another map update.
So shuttles are still acting whacky, requires an admin to teleport to
the supply shuttle and manually add it to the console in space to the
west on z-level 2, after that things should work proper.

Added a bunch of changes to the initial shuttle, different vending
machines/lockers for miners etc etc, all suggestions from players. Also
removing some of the garbage files that just bloat up file size.